[PIR] split TestSundryAPIStatic (#62909)
